Implement AVG aggregate function. Fix some issues with null values and arithmetic operators (division in particular). Move some utility methods to linq4j.

+++ src/test/java/net/hydromatic/optiq/test/SqlFunctionsTest.java
@@ -32,7 +32,7 @@ import static net.hydromatic.optiq.runtime.SqlFunctions.*;
  */
 public class SqlFunctionsTest extends TestCase {
     public void testCharLength() {
-        assertEquals((Integer) 3, charLength("xyz"));
+        assertEquals(3, charLength("xyz"));
     }
 
     public void testConcat() {
